Load data once and share via context
