+++ components/sections/ArticlesSection.tsx
@@ -1,11 +1,10 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import { Link } from 'react-router-dom';
-import { ARTICLES_DATA } from '../../constants';
 import { Article } from '../../types'; // Ensure Article type is imported
 import AnimatedDiv from '../ui/AnimatedDiv';
 import Button from '../ui/Button';
 import { CalendarDays, UserCircle, Tag, ChevronLeft } from 'lucide-react';
-import { supabase } from '../../utils/supabaseClient'; // Import supabase client
+import { useData } from '../../contexts/DataContext';
 
 interface ArticleCardProps {
     article: Article;
@@ -79,68 +78,8 @@ interface ArticlesSectionProps {
 }
 
 const ArticlesSection: React.FC<ArticlesSectionProps> = ({ maxItems, showTitle = true }) => {
-    const [allArticles, setAllArticles] = useState<Article[]>(ARTICLES_DATA);
-    const [error, setError] = useState<string | null>(null);
-
-    useEffect(() => {
-        const fetchArticles = async () => {
-            try {
-                // Assuming 'articles' is your table name in Supabase
-                // and columns match the Article type.
-                // Ensure Supabase returns `id` as string or cast it.
-                // Ensure date format is consistent or parse/format it.
-                const { data: supabaseArticles, error: supabaseError } = await supabase
-                    .from('articles') // Make sure this table name is correct
-                    .select('*'); // Select all columns, adjust if needed
-
-                if (supabaseError) {
-                    console.error('Error fetching articles from Supabase:', supabaseError);
-                    setError('Failed to load articles from database. Error: ' + supabaseError.message);
-                    // Keep existing articles if fetch fails
-                    setAllArticles(ARTICLES_DATA);
-                    return;
-                }
-
-                if (supabaseArticles) {
-                    // Transform Supabase data to match our Article interface
-                    const fetchedArticles: Article[] = supabaseArticles.map((supaArticle: any) => {
-                        const created = supaArticle.date || supaArticle.created_at;
-                        const bodyText: string = supaArticle.fullContent || supaArticle.body || '';
-
-                        return {
-                            ...supaArticle,
-                            fullContent: bodyText,
-                            excerpt: supaArticle.excerpt || bodyText.substring(0, 150),
-                            author: supaArticle.author || 'צוות מכון אביב',
-                            imageUrl: supaArticle.imageUrl,
-                            date: created ? new Date(created).toLocaleDateString('he-IL', { day: '2-digit', month: '2-digit', year: 'numeric' }) : 'N/A',
-                            id: String(supaArticle.id),
-                        } as Article;
-                    });
-
-                    // Combine and remove duplicates (preferring Supabase articles if IDs clash)
-                    const combinedArticles = [
-                        ...fetchedArticles,
-                        ...ARTICLES_DATA.filter(localArticle =>
-                            !fetchedArticles.find(fetched => String(fetched.id) === String(localArticle.id))
-                        )
-                    ];
-                    setAllArticles(combinedArticles);
-                } else {
-                    // No articles from Supabase, just use local ones
-                    setAllArticles(ARTICLES_DATA);
-                }
-            } catch (err) {
-                console.error('Error in fetchArticles:', err);
-                setError('An unexpected error occurred while fetching articles.');
-                setAllArticles(ARTICLES_DATA); // Fallback to local data
-            }
-        };
-
-        fetchArticles();
-    }, []);
-
-    const articlesToDisplay = maxItems ? allArticles.slice(0, maxItems) : allArticles;
+    const { articles, error } = useData();
+    const articlesToDisplay = maxItems ? articles.slice(0, maxItems) : articles;
 
     return (
         <section className="py-16 sm:py-20 md:py-24 bg-slate-100 dark:bg-slate-900">
@@ -172,7 +111,7 @@ const ArticlesSection: React.FC<ArticlesSectionProps> = ({ maxItems, showTitle =
                     ))}
                 </div>
 
-                {maxItems && allArticles.length > maxItems && (
+                {maxItems && articles.length > maxItems && (
                     <AnimatedDiv animation="fadeInUp" delay={0.5} className="text-center mt-12 sm:mt-16">
                         <Button href="/articles" variant="primary" size="lg">
                             לכל המאמרים
